Pedidos: Mudança do Id de Integer para long e métodos de Delete pedidos, AddItem e Delete Item

+++ src/main/java/com/lipcam/PedidosApiSpring/dtos/pedidos/PedidosItensDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class PedidosItensDTO {
-    private  Integer idProduto;
+    private  Long idProduto;
     private Integer quantidade;
 }
