+++ src/main/java/com/lipcam/PedidosApiSpring/repositories/PedidosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PedidosRepository extends JpaRepository<Pedidos, Integer> {
+public interface PedidosRepository extends JpaRepository<Pedidos, Long> {
     List<Pedidos> findByClientes(Clientes clientes);
 }
